\name{cdna}
\docType{data}
\alias{cdna}
\title{Example cDNA data}

\description{cDNA data from a one-colour experiment from the Stanford
	Microarray Database. The data set is stored as an intensity matrix
	with 42624 genes (rows) and 4 replicates (columns).

	The experiment numbers are 40430, 40571, 34905 and 34912.
}
\usage{
 data(cdna)
}
\format{A 42624x4 matrix}
\source{Stanford Microarray Database}
\references{
McCaffrey, R.L., Fawcett, P., O'Riordan, M. Lee, K., Havell, E.A. Brown, P.O. and Portnoy, D.A. (2004) A specific gene
expression program triggered by Gram-positive bacteria in the cytocol, Proc. Nat. Acad. of Science, 101, 11386-11391
}
\keyword{datasets}
