\name{function.from.vector}
\alias{function.from.vector}
\title{Function applied for the computation of the Data-Driven Haar-Fisz transform}
\description{Applies the function which.min.diff to the argument.vect vector
	with the argument x}
\usage{
function.from.vector(x, y, argument.vect)
}
\arguments{
  \item{x}{a vector of data in ascending order}
  \item{y}{a vector with length(y) = length(x)}
  \item{argument.vect}{can be a vector}
}
\value{
  \item{function.from.vector}{help function for data transformation}
  }
\author{Piotr Fryzlewicz <p.fryzlewicz@imperial.ac.uk>}
\keyword{methods}
