\name{KURTOSIS}
\alias{KURTOSIS}
\title{Kurtosis Coefficient estimator}
\description{
  Estimates the kurtosis coefficient of a data vector  
}
\usage{
KURTOSIS(x)
}
\arguments{
  \item{x}{the data vector}
}
\value{The estimated kurtosis of x }
\author{Efthimios Motakis <e.motakis@bris.ac.uk}
\keyword{univar}
