\name{MOMENTS}
\alias{MOMENTS}
\title{Moment estimator}
\description{
	Computes the rth central moment of the data x
}
\usage{
MOMENTS(x, r)
}
\arguments{
  \item{x}{the data vector}
  \item{r}{parameter that describes which moment will be produced}
}
\value{The rth central moment of x}
\author{Efthimios Motakis <e.motakis@bris.ac.uk.}
\keyword{documentation}
