\name{ShipCreditFlow}
\docType{data}
\alias{ShipCreditFlow}
\title{Example Shipping credit flow data}

\description{Shipping credit flow data. A set of monthly credit
	transaction flow series recorded by the Central Bank of Cyprus.
	The time range of the data is January 2002 to December 2014 and
	contains 156 observations.
}
\usage{
 data(ShipCreditFlow)
}
\format{A vector of length 156}
\source{Central Bank of Cyprus}
\references{
Michis, A. A. and Nason, G.P. (2016)
Case Study: Shipping trend estimation and prediction via multiscale
variance stabilisation. \emph{Journal of Applied Statistics} (to appear)
}
\keyword{datasets}
